+++ packages/ethereum/contracts/src/HoprNetworkRegistry.sol
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: GPL-3.0
 pragma solidity ^0.8.0;
 
-import '@openzeppelin/contracts/access/Ownable.sol';
+import 'openzeppelin-contracts-4.8.3/contracts/access/Ownable.sol';
 import './IHoprNetworkRegistryRequirement.sol';
 
 /**